src/workerPool.ts
import { AsyncResource } from 'node:async_hooks';
import { EventEmitter } from 'node:events';
import { Worker } from 'node:worker_threads';
import { WorkerOptions } from "worker_threads";

const kTaskInfo = Symbol('kTaskInfo');
const kWorkerFreedEvent = Symbol('kWorkerFreedEvent');

class WorkerPoolTaskInfo extends AsyncResource {
  private readonly callback: () => void;

  constructor(callback) {
    super('WorkerPoolTaskInfo');
    this.callback = callback;
  }

  done(err, result) {
    this.runInAsyncScope(this.callback, null, err, result);
    this.emitDestroy();  // `TaskInfo`s are used only once.
  }
}

class WorkerPool extends EventEmitter {
  private readonly workers: Array<Worker>;
  private freeWorkers: Array<Worker>;

  constructor(numThreads: number, workerFile: string, options?: WorkerOptions) {
    super();
    this.workers = [];
    this.freeWorkers = [];

    for (let i = 0; i < numThreads; i++) {
      this.addNewWorker(workerFile, options);
    }
    console.log(`Created ${numThreads} workers`);
  }

  addNewWorker(workerFile: string, options?: WorkerOptions) {
    const worker = new Worker(workerFile, options);
    worker.on('message', (result) => {
      // In case of success: Call the callback that was passed to `runTask`,
      // remove the `TaskInfo` associated with the Worker, and mark it as free
      // again.
      worker[kTaskInfo].done(null, result);
      worker[kTaskInfo] = null;
      this.freeWorkers.push(worker);
      this.emit(kWorkerFreedEvent);
    });
    worker.on('error', (err) => {
      // In case of an uncaught exception: Call the callback that was passed to
      // `runTask` with the error.
      if (worker[kTaskInfo])
        worker[kTaskInfo].done(err, null);
      else
        this.emit('error', err);
      // Remove the worker from the list and start a new Worker to replace the
      // current one.
      this.workers.splice(this.workers.indexOf(worker), 1);
      this.addNewWorker(workerFile, options);
    });
    this.workers.push(worker);
    this.freeWorkers.push(worker);
    this.emit(kWorkerFreedEvent);
  }

  runTask(task, callback) {
    console.log('this.freeWorkers.length : ', this.freeWorkers.length);
    if (this.freeWorkers.length === 0) {
      // No free threads, wait until a worker thread becomes free.
      this.once(kWorkerFreedEvent, () => this.runTask(task, callback));
      return;
    }

    const worker = this.freeWorkers.pop();
    worker[kTaskInfo] = new WorkerPoolTaskInfo(callback);
    worker.postMessage(task);
  }

  async close() {
    await Promise.all(this.workers.map(worker => {
      worker.unref();
      return worker.terminate()
    }));
  }
}

export default WorkerPool;
